+++ changelog.md
@@ -1,3 +1,11 @@
+# v1.1.0
+
+ * Updated for 2.2074
+ * Fixed relaxed mode
+ * Added more filtered words
+ * Added word replacements
+ * Added filtering levels
+
 # v1.0.2
 
  * Updated for 2.206